Removing hard-coded list
Replaced hard-coded numbers of spaces with the equivalent dynamic code.
Added an EditorConfig

My preferences are set, but can be updated.
Replaced parameter "text" with "input_text"

This makes it a bit clearer to people reading the code (who are perhaps used to other languages) that the variable is a parameter, not a class